pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.10, <3.14"

dependencies = [
    "livekit-agents[turn-detector]~=1.3",
    "livekit-plugins-elevenlabs~=1.3",
    "livekit-plugins-openai~=1.3",
    "livekit-plugins-cartesia~=1.3",
    "livekit-plugins-silero~=1.3",
    "ten-vad @ git+https://github.com/TEN-framework/ten-vad.git",
    "python-dotenv",
    "soxr",
    "torch>=2.0,<2.6",
    "torchaudio>=2.0,<2.6",
    "pipecat-ai[local-smart-turn-v3]>=0.0.102,<0.1",
    "einops",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"
extend-exclude = ["src/vendor"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "platform_system == 'Linux' and platform_machine == 'x86_64'" },
]
torchaudio = [
    { index = "pytorch-cpu", marker = "platform_system == 'Linux' and platform_machine == 'x86_64'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
